Fix Error in asynchcommand.vim using [t]csh
The bug is because &shellredir can be “>”, “>&”, or “>%s 2>&1” according
to :help shellredir. Using Bash, “>%s 2>&1” is used and the printf
statement in the plugin can interpolate the filename into %s. In C
Shell, “>&” is returned and the printf fails because “too many
arguments” were specified.

To simulate the problem you can do a :set shellredir=>& and then
exercise the plugin.

Thanks to Chad Moore.

Bug posted at:
    https://github.com/pydave/daveconfig/issues/2