Pass the return code to the handler
Grab the return code with the appropriate variable and pass it to the
handler. If you do any nesting like tab_restore, then you need to copy
the return code to your child env.

None of the handlers use the return code, but the documentation has a
working example. (I test the ant build example by making an executable
file called ant that cats an existing or nonexisting file -- depending
on the test case.)

Tested on ubuntu, but not Windows.