Fix #13: Add requirement check to plugin
AsyncCommand requires clientserver. This check is already done in the
autoload files, so bring it to the plugin too and add an error message.

Also check we're nocompatible since the autoload check for that. (We
also save cpo, but that's to prevent the user's cpoptions from
interfering with regular operation.)