Rephrase comment
I think I understand this better. Any time you run a command in console
vim, it flips to your terminal so you can see the output of the command
AsyncCommand doens't show anything, but that flip still happens and
vim's UI isn't redrawn properly (because of lazyredraw?). So force the
redraw.