Fix #15: Try harder to find current vim
Use the config flags for the current executable to figure out what
version of vim you're running and try to execute that. Vim still needs
to be in your PATH (or else you can set g:asynccommand_prg).

Give an error message and abort if Vim can't be found.

Also only use mvim if the gui is actually running. Looks like that's the
right call from this issue on janus:
    https://github.com/carlhuda/janus/pull/372