Fix #18: Return empty string from done()
While AsyncCommand uses remote-expr, we don't actually want any output
to be displayed, so always return an empty string.

As moorec commented on #9, it's possible for this function to output a 0
and for that to make it on the screen. Presumably that's not a problem
with vim because it's consistent enough about redrawing the screen. Gvim
doesn't let the output slip by.