Add a pending command listing
Add AsyncPending to make it possible to find in-progress log files.
Helpful to see the output of a pending compile command or to figure out
what you have running in the background.

Only uses existing information, and only some handlers add a title to
their env Dictionary. May be more useful if there's more information
here (and it might be nice for all of the output windows to have useful
titles).