Add support for AsyncShell! for readonly output
AsyncShell creates a split, but it would be nice to map `q` to close the
buffer just like :Verbose and :Gstatus. Create banged version of
AsyncPending that follow their convention and make it a preview window,
using :bdelete to close the window. Also set nomodifiable since if you
can't use `q`, then you probably cannot modify.

Aside from 'previewheight', further customization (using a different
kind of buffer) is limited to implementing your own version of
asynchandler#split.

Does not make the buffer 'readonly' because that seems to be for files
that you cannot write to. There's no problem writing the file (it's
a temp file), I just want to limit modification.